requirements.txt
Django==1.5.4
Markdown==2.5.2
Pillow==2.6.1
argparse==1.2.1
decorator==3.4.0
django-filter==0.9.0
django-social-auth==0.7.28
httplib2==0.9
lucene-querybuilder==0.2
-e git+https://github.com/scholrly/neo4django/@68202bba775713172fe92e95f87e43520f039104#egg=neo4django-master
neo4jdb==0.0.8
oauth2==1.5.211
py2neo==1.6.4
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2014.7
reportlab==3.1.8
requests==2.4.3
six==1.8.0
svglib==0.6.3
wsgiref==0.1.2
